mark endpoint as internal in configuraiton-service

+++ configuration-service/restapi/operations/service_resource/get_project_project_name_stage_stage_name_service_service_name_resource_parameters.go
@@ -117,6 +117,7 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) Bi
 	if err := o.bindStageName(rStageName, rhkStageName, route.Formats); err != nil {
 		res = append(res, err)
 	}
+
 	if len(res) > 0 {
 		return errors.CompositeValidationError(res...)
 	}
@@ -132,7 +133,6 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: false
 	// AllowEmptyValue: false
-
 	if raw == "" { // empty values pass all other validations
 		// Default values have been previously initialized by NewGetProjectProjectNameStageStageNameServiceServiceNameResourceParams()
 		return nil
@@ -156,10 +156,10 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: false
 	// AllowEmptyValue: false
-
 	if raw == "" { // empty values pass all other validations
 		return nil
 	}
+
 	o.NextPageKey = &raw
 
 	return nil
@@ -174,7 +174,6 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: false
 	// AllowEmptyValue: false
-
 	if raw == "" { // empty values pass all other validations
 		// Default values have been previously initialized by NewGetProjectProjectNameStageStageNameServiceServiceNameResourceParams()
 		return nil
@@ -196,11 +195,11 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 // validatePageSize carries on validations for parameter PageSize
 func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) validatePageSize(formats strfmt.Registry) error {
 
-	if err := validate.MinimumInt("pageSize", "query", *o.PageSize, 1, false); err != nil {
+	if err := validate.MinimumInt("pageSize", "query", int64(*o.PageSize), 1, false); err != nil {
 		return err
 	}
 
-	if err := validate.MaximumInt("pageSize", "query", *o.PageSize, 50, false); err != nil {
+	if err := validate.MaximumInt("pageSize", "query", int64(*o.PageSize), 50, false); err != nil {
 		return err
 	}
 
@@ -216,6 +215,7 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: true
 	// Parameter is provided by construction from the route
+
 	o.ProjectName = raw
 
 	return nil
@@ -230,6 +230,7 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: true
 	// Parameter is provided by construction from the route
+
 	o.ServiceName = raw
 
 	return nil
@@ -244,6 +245,7 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: true
 	// Parameter is provided by construction from the route
+
 	o.StageName = raw
 
 	return nil
